cmake: delete FIXME for libssh2 detection
configure has an interesting default setting for dependencies
where detection never match up with cmake:

It's called 'off' and means to look for the lib at 'standard
locations'. Which possibly means system locations. It misses
most dependencies on macOS, because just a couple are shipping
with the OS at system locations.